requirements.txt
web3==5.26.0
Telethon==1.24.0
python-decouple==3.6
requests==2.26.0
